wise-ft-clip/requirements.txt
captum==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.0
ftfy==6.1.3
humanfriendly==10.0
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
onnx==1.15.0
onnxruntime==1.16.3
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sympy==1.12
tensorboard==2.16.2
tensorflow-io[tensorflow]==0.37.1
threadpoolctl==3.2.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12